requirements.txt
viam-sdk=0.5.0